refactor(prompts): delegate tool implementations to custom_tools modules
- Created custom_tools directory with specialized implementation files- Refactored tools.py to act as facade delegating to custom implementations- Added comprehensive README for custom_tools explaining architecture- Updated main prompts README with new structure and best practices- Improved documentation for semantic search and collaboration features- Enhanced tools to leverage the new semantic search capabilities in registry
